pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "elasticsearch>=9.0.0",
    "flask-session>=0.8.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "opensearch-py>=2.8.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "stripe>=12.0.1",
]
